+++ frontend/components/app/cards/card/tasks/task.vue
@@ -36,6 +36,7 @@
 <script type="text/babel">
     import taskService from '../../../../../services/task'
     import Utils from '../../../../../services/utils'
+    import eventEmitter from '../../../../../services/event-emitter'
     import $ from 'jquery'
 
     // noinspection JSUnusedGlobalSymbols
@@ -61,7 +62,7 @@
             },
             delete() {
                 taskService.delete(this.model.id).then(() => {
-                    this.$dispatch('tasks.-', this.model)
+                    eventEmitter.emit('tasks.deleted', this.model)
                 })
             },
             edit() {
@@ -77,7 +78,7 @@
                 Utils.parseTask(this.model)
 
                 taskService.update(this.model.id, this.model).then(() => {
-                    this.$dispatch('tasks.updated')
+                    eventEmitter.emit('tasks.updated')
                 })
             },
             showWarning() {