+++ frontend/components/app/cards/menu.vue
@@ -3,7 +3,7 @@
         <a class="item"
            v-link="{ name: 'card', params: { id: model.id }, activeClass: 'active' }"
            v-for="model in models">
-            {{ model.name }}
+            {{ model.name }} <div class="ui teal label">{{ getCount(model.id) }}</div>
         </a>
         <add :models="models"></add>
     </div>
@@ -12,18 +12,31 @@
 
 <script type="text/babel">
     import AddInput from './menu/add'
+    import eventEmitter from '../../../services/event-emitter'
 
     export default {
 
-        props:    ['models'],
+        props:    ['tasksCount', 'models'],
 
         components: {
             add: AddInput
         },
 
+        methods: {
+            getCount(id) {
+                const taskCount = this.tasksCount.find(count => count.card === id)
+                return taskCount === undefined ? 0 : taskCount.count
+            }
+        },
+
         init() {
-            this.$on('cards.+', (model) => {
+            eventEmitter.on('cards.created', (model) => {
+                // add card to models
                 this.models.push(model)
+            })
+            eventEmitter.on('tasks.created', model => {
+                // increment tasks count
+                this.tasksCount.find(count => count.card === model.cardId).count++
             })
         }