+++ frontend/services/event-emitter.js
@@ -1,0 +1,4 @@
+import EventEmitter from 'eventemitter2'
+
+const emitter = new EventEmitter()
+export default emitter