+++ frontend/services/task.js
@@ -4,6 +4,26 @@ import Promise from 'promise'
 
 export default {
 
+    countByCards() {
+        return new Promise((resolve, reject) => {
+            Vue.http.get('api/index.php/cards/tasks/count').then(response => {
+                resolve(response.data)
+            }, error => {
+                reject(error)
+            })
+        })
+    },
+
+    delete(id) {
+        return new Promise((resolve, reject) => {
+            Vue.http.delete(`api/index.php/tasks/${id}`).then(response => {
+                resolve(response.data)
+            }, error => {
+                reject(error)
+            })
+        })
+    },
+
     queryByCardId(id) {
         return new Promise((resolve, reject) => {
             Vue.http.get(`api/index.php/cards/${id}/tasks`).then(response => {
@@ -17,16 +37,6 @@ export default {
     save(cardId, model) {
         return new Promise((resolve, reject) => {
             Vue.http.post(`api/index.php/cards/${cardId}/tasks`, model).then(response => {
-                resolve(response.data)
-            }, error => {
-                reject(error)
-            })
-        })
-    },
-
-    delete(id) {
-        return new Promise((resolve, reject) => {
-            Vue.http.delete(`api/index.php/tasks/${id}`).then(response => {
                 resolve(response.data)
             }, error => {
                 reject(error)